data/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
audioread==3.0.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.13
fastapi-cli==0.0.7
fonttools==4.58.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
msgpack==1.1.1
numba==0.61.2
numpy==2.2.6
packaging==25.0
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.4
rich==14.0.0
rich-toolkit==0.14.7
scikit-learn==1.7.0
scipy==1.16.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
starlette==0.46.2
threadpoolctl==3.6.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
redis
apscheduler
